+++ src/lib/errors/index.ts
@@ -1,6 +1,10 @@
-export {RequestError, UnexpectedHttpStatusError} from "./http";
-export {WrongCredentialsError} from "./wrong-credentials";
-export {WrongCredentialsLimitError} from "./wrong-credentials-limit";
+export { RequestError, UnexpectedHttpStatusError } from "./http";
+export { WrongCredentialsError } from "./wrong-credentials";
+export { WrongCredentialsLimitError } from "./wrong-credentials-limit";
 
+import * as endpointRegistrationError from "./endpoint-registration";
+import * as LoginRateLimitExceeded from "./login-rate-limit-exceeded";
 import * as microsoftAccount from "./microsoft-account";
-export {microsoftAccount};
+import * as RedirectionLimit from "./redirection-limit";
+
+export { microsoftAccount, endpointRegistrationError, LoginRateLimitExceeded, RedirectionLimit };